requirements/requirementsWorker.txt
arq
asyncpg
sqlalchemy
pandas
pydantic
fastapi
fastapi-users[sqlalchemy]
PyPDF2
python-docx
openpyxl 
transformers
aioboto3
pydub
unstructured
opencv-python-headless
pillow
dill
ultralytics==8.2.16
lapx>=0.5.2